Alot has changed, there will probably be bugs present in this version:
Removed "class" field from signup.

On signup the user will have "status" set to 0, the admin has to assign each user a class by posting to /admin/users/assign, 2 fields are required in this post; "class" and "users" which an array of user ID's.

Added field "course" to users

Exercises and answers no longer have the field "class" but will use "course" instead.

Exercises now have a field "revealdate", if the current date is before the revealdate, /users/exercises won't list the exercise. If an admin receives the list of exercises (/admin/exercises) there will be a field named "revealed" which is set to true, maybe make it's tile gray when it's false? @ErazerBrecht